Fix issues #6226 and #6227 in multiline-enabled TextInput  (#6249)
* Add TextInput._visible_lines_range attr (+minor fix)

TextInput._visible_lines_range attribute contains the range of lines
which were partially (at least 1px width line) or fully rendered in the
TextInput's viewport during the last _update_graphics() call.
The attribute is meant to be mainly used in TextInput's scrolling
unittests.
The minor fix makes _update_graphics() method stop rendering the
lines which share no pixels with current viewport and interrupt the for
loop after the last visible line.

* Draw cursor only inside the TextInput's viewport

This commit implements drawing the cursor correctly on the border
between the padding and actual viewport of the TextInput widget. The
cursor is fully drawn inside the viewport, partially drawn on the border
and not drawn outside the viewport.

* Fix scrolling with ctrl+cursor_up, ctrl+cursor_down

As the TextInput.do_cursor_movement documentation states the
"ctrl+cursor_up" and "ctrl+cursor_down" combinations should scroll text
one line up or down respectivelly without changing the cursor position.
Also the text should be scrolled both when the cursor is inside the
viewport and outside of it.
These 2 conditions were broken and are fixed by this commit.

* Don't change cursor position when scrolling TextInput

* Fix TextInput's viewport freeze on scroll_up

TextInput.on_touch_down was checking the possibility to scroll_up
using the positions of the rendered lines' rects. These positions
are not updated when the lines are skipped (e.g. during fast scroll
or ctrl+cursor_home) which lead to scroll freeze

* Reduce code duplication in TextInputGraphicTest