window: sdl2: use GLES context when ES2 is forced
On the desktop, multiple context types are generally available,
including full OpenGL with core and compatibility contexts, in addition
to ES contexts.

Certain extensions, such as GL_OES_EGL_image_external require a GLES
context, but SDL2 defaults to a core or compatibility context when
available.

This patch forces a GLES context when ES2 is forced using KIVY_GRAPHICS.