No need to import Properties!
We add all properties to a Widgets dict in __prepare__ with a ChainMap.  This allows, in the context of a Widget class def, to avoid importing any properties at all -- as property imports can get rather lengthy.  We also added a proof-of-concept auto-loader of kv lang which could be provided either in a special class attribute, `__KV__`, or a file in the same directory with the same name (in lowercase) as the widget.