graphics: fix ignored alpha value in hsv mode
When using the Color instruction in hsv mode, the alpha value is
initialized as 1.0, no matter what value was defined by the user. This
bug was introduced trying to fix a similar issue [1].

[1]: https://github.com/kivy/kivy/pull/6232

To fix it, I set the alpha value to what the user provides, or 1.0 if it
is not defined.

Issue related:

https://github.com/kivy/kivy/issues/6590