Accept numeric instances in NumericProperty
NumericProperty class only accepts int, float, long, tuple and list
types. Because of that, numeric instances like numpy's float64 raise a
ValueError exception. For example, this code will crash (it is
simplified to focus on the error only):

```
import numpy as np
from kivy.uix.widget import Widget
from kivy.properties import NumericProperty


class CrashWidget(Widget):
    array_of_numbers = np.array([1.0])
    numeric_value = NumericProperty(array_of_numbers[0])


if __name__ == "__main__":
    CrashWidget()
```

For the NumericProperty class, array_of_numbers[0] is not a valid type
because it is not one of int, float, long, tuple or list. However, it is
an instance of float (`isinstance(array_of_numbers[0], float)` is True).
That said, the proposed solution is to check for the instance of the
accepted NumericProperty types instead of checking for specific types
only.

Related issue:

https://github.com/kivy/kivy/issues/6982