Horizontal scrolling disabled if no overflow
When no overflow, scrolling gave division by zero error. This was because the width of scroll bar became 1, and that multiplied by window width became equal to window width. Subtraction of the two gave zero, where Python gave an error. A simple if statement is added so that scrolling is disabled if scroll bar width is 1.