Fix #7214 Inspector broken
This was caused by a circular import
- When the window is created, it registers the module
- The module imports TextInput
- TextInput imports the Window object, to bind to it
- But the window is still being created, so the module tries to create
  it again.
- Python detects the circular import, and returns an incomplete
  inspector module, without a start method
- the module can't be started, and the Window object is fubar.

Solution.
The TextInput import being the cause of the problem, instead of
importing it, we can import the Factory and lookup the TextInput object
from it.
To avoid similar problems potentially happening in the future, the other
widgets and instructions have also been converted to using the Factory.

This might still be brittle, and possibly changing the way the modules
are registered would be more future-proof.