Window: prevent "empty" mousewheel event from breaking scrollview
It seems on some machine, especially lenovo ones, but maybe others,
scrolling with the touchpad can create a lot of "fake" scrolling events,
without any associated offsets.

The way kivy works before this change is to dispatch these events as
"mousewheelup" (which indicates a scroll down). This makes it really
hard to scroll up in this situation, as there is a constant stream of
events pushing in the other direction, and the more we try to scroll,
the more there are.

This change ignore these empty events.