Ensure tuple value for Window.clearcolor.
The Window clearcolor property documentation (see kivy.core.window, line 473)
is declared as a tuple (and not as a list).

When one want e.g. replace the alpha value with a expression like
`Window.clearcolor[:3] + (0.5, )` an exception get raised if
Window.clearcolor contains a list object.

Would like to make this consistent to prevent extra/ugly type
conversion code for expressions like this, because all other
places in the kivy code base (apart from the two in this
PR) are using tuples for Window.clearcolor.