FIX: Disable `KIVY_NO_ARGS` depending on its value
This fix enables the environment value `KIVY_NO_ARGS` to be present but still
not disabling kivy cli parsing.

Before this, kivy only checked for `KIVY_NO_ARGS` environment variable to be
existent, wihout checking its value. Setting it to a "falsy" value (eg 0,
"false", etc) would still disable cli arg parsing, which is could be
misleading.

With this fix, only setting it to one of ('true', '1', 'yes') actually
disables it.