Fixes video thread that sleep before dispatching eos.
Sleeping here was to avoid a busyloop when waiting the user/code to do something when eos happened.
I am not trying to refactor this code, but this Event allow a instant reactivity from the mainthread to the frame's thread, and completly cut the lag, visually. No visual jump anymore on perfect-loop video.

Fixes #7417