Reformulation of the Switch to allow better control over its properties.
The properties used to compose the Switch, such as the background image and the slide button image previously unavailable, are now accessible.
Other properties have been added to allow greater control over the dimensions of the Switch components. The relative properties of the size hint, for example, allow relative control over the dimensions of both the background image and the image of the slider, in relation to the Switch widget itself.