Re-implements the Bubble widget.
The old Bubble widget design was complicated and used a GridLayout with
multiple dummy widgets solely to get an arrow placed as desired.
Multiple magic numbers like 99 columns or 99 rows were undocumented
and unintuitive. Probably some of these numbers even add unnecessary
limitations to the Bubble widget.
The Bubble widget was fused with the 'BubbleContent' which by design
makes the Bubble widget feel very unflexible.
Sizing the Bubble according to the content was unintuitive if not even
impossible for some cases and the arrow positioning was relying on magic
numbers as well. Lastly, the old implementation lacked test coverage
with very few tests that did not even cover the core logic of the
widget.

The new implementation reduces the core logic (formerly "on_arrow_pos",
~100 lines of code) to ~30 lines plus a well documented lookup structure
without magic but intuitive values. Furthermore, it adds proper inline
documentation for maintainers.

Furthermore, completely flexible arrow positioning was implemented and
allows free positioning of the arrow to all sides of the content by
selecting a (relative) position on the Bubble widget border.

The BubbleContent widget is a seperate styled BoxLayout and the code and
documentation makes clear that any other Widet or Layout can be placed as
content. The background_color, border and border_auto_scale properties
have been attached to the BubbleContent instead of the Bubble itself
which removes the unnecessary connections between the Bubble itself and it's
BubbleContent.

Properties, that refer to multiple dimensions are cleanly integrated
through ReferenceListProperties.

The rework of the Bubble widget fixes the issue #4348 which had been
opened in 2016 and remained unsolved.

Minimal changes are necessary for existing applications using the Bubble
widget:
- A BubbleContent (can be any Layout or a single Widget) is added as only child
  to the Bubble (similarly to the ActionBar/ActionView widgets).
- The background_color, border and border_auto_scale properties have
  been removed from the Bubble as they belong to the BubbleContent

Signed-off-by: AnthonyZimmermann <anthony.zimmermann@protonmail.com>