Implement TextInputSpinner
Implement a TextInputSpinner class which basically is a spinner with
a TextInput instead of a button. This allows for user input of values
that are not in the set of predefined options.
Furthermore, a filtering function has been implemented which filters the
options based on the current input. The implementation is flexible and
allows developers to implement their custom filter function.
Implements tests for the filtering feature.