Highlight parenthetical expression with hyphens
The expression "and source code files for each step" was only partially
in a parenthesis: there wasn't a closing comma after the word "step" to
highlight that the expression was extra information in the sentence.
This commit changes the punctuation around the parenthetical expression
to use hyphpens to make it stand out more and adds the missing closing
punctuation.