Use consistent source code notes in pong tutorial (#7929)
* Highlight parenthetical expression with hyphens

The expression "and source code files for each step" was only partially
in a parenthesis: there wasn't a closing comma after the word "step" to
highlight that the expression was extra information in the sentence.
This commit changes the punctuation around the parenthetical expression
to use hyphpens to make it stand out more and adds the missing closing
punctuation.

* Add missing full stop at end of note sentence

Since this note is a sentence, it probably should have a full-stop at
the end.

* Mention source location consistently in pong tutorial

The two notes mentioning the source code location were formatted
differently.  To be consistent, they both now use the same formatting.