initialize our logger inside of kivy/__init__.py
The logger can end up not being `_configured` if this is not called.
This can occure when log level is set to a stricter level like `error`
and no errors occur, then when we call the purge_logs function our
handler could be un-configured and not have a log_dir attribute --
resulting in the function bailing out.