Use SDL_GetDisplayDPI as fallback method to get DPI
Use [SDL_GetDisplayDPI](https://wiki.libsdl.org/SDL2/SDL_GetDisplayDPI) to get the display DPI if no high DPI was detected using the method described in the warning for
the SDL_GetDisplayDPI function (which is the current method for guessing the DPI).

If the X server is configured with the DPI of the display then this GetDisplayDPI method
returns the correct DPI on Linux.