Scale metrics density for linux
If higher than normal DPI is configured using `KIVY_DPI` or other means then the density of the screen is higher and this should be reflected in the density.

This will make the density calculation on linux the same as Windows and macOS.

Unit test change:

Use almost equal in density property unit tests
Since the density is a floating point number, the tests can fail
if the calculated value is slightly rounded.