Fixing `TypeError` on `parse_level` method
When you are write something like this:
```yml
BoxLayout:
    orientation: 'vertical'
        Button:
```
You would get this error:
```
  File "/.venv/lib/python3.10/site-packages/kivy/lang/parser.py", line 805, in parse_level
    if current_property[:3] == "on_":
TypeError: 'NoneType' object is not subscriptable
```
Because `current_property` is None, and on [line 756](https://github.com/kivy/kivy/blob/master/kivy/lang/parser.py#L756) we try to subscript it.
But this error is not informative and does not allow the developer to know where the error comes from.
In this case we should show in which line there is a invalid declaration,  by raising a `ParserException`. 
It works.