Copy edits: Standardize on spelling
Changed comments (only) to use American spellings of common words.

(As an Australian, it hurts me to use the American spelling, but it looks more professional to have one standard throughout, and the code itself uses American spellings, so I kept it consistent.)

Also trivial copy-edits to comments in RecycleView.