fix pythonw and pyinstaller issue
Kivy apps will crash if run from pythonw or if they are built with pyinstaller 5.7  In both cases sys.stderr is set to None, causing a recursion error in the python logger.  This fix sets the KIVY_NO_CONSOLELOG environment if sys.stderr is None.  This preserves the behavior prior to pyinstaller 5.7, and allows correct operation under pythonw.exe