Prevent apps from crashing when `sys.stderr` is `None` (pythonw and pyinstaller 5.7) (#8345)
* fix pythonw and pyinstaller issue

Kivy apps will crash if run from pythonw or if they are built with pyinstaller 5.7  In both cases sys.stderr is set to None, causing a recursion error in the python logger.  This fix sets the KIVY_NO_CONSOLELOG environment if sys.stderr is None.  This preserves the behavior prior to pyinstaller 5.7, and allows correct operation under pythonw.exe

* Update logger.py

Moved changed to the add_kivy_handlers() function

* modified per @pythonic64

* Added test for fix of issue 8345

* Lines shortened for pep8 compliance

* Removed "\" for pep8 compliance

* Used parens to shorten lines to be consistent with the rest of the file, for pep8 compliance

* Used "\" for line continuation, for pep8 compliance

* Update kivy/tests/test_logger.py

Co-authored-by: Mirko Galimberti <me@mirkogalimberti.com>

---------

Co-authored-by: Mirko Galimberti <me@mirkogalimberti.com>