Update _text_pango.pyx
This is just a typo fix for the line `global kiby_fontfamily_cache`, the kiby instead of kivy being the typo.
I don't use Pango myself and it seems a little odd that this typo is in the code when it looks like it might break something? Someone may want to have a look into this.