Fixes avplayer seek and state on iOS (not tested on OSX)
The previous implementation was using playAtTime_, but 100% sure it was
not working as playAtTime_ schedules the audio to play at a specific
time, not to seek within the current audio file. To seek to a specific
position in the audio, we should use the currentTime property of
AVAudioPlayer.

In addition, i added a delegate to correctly catch when the sound is
stopping when reaching eof, which make the state property correctly
work.