`avplayer` audio provider: Fixes `seek` and `state` on iOS and macOS (#8741)
The previous implementation was using playAtTime_, but 100% sure it was
not working as playAtTime_ schedules the audio to play at a specific
time, not to seek within the current audio file. To seek to a specific
position in the audio, we should use the currentTime property of
AVAudioPlayer.

In addition, i added a delegate to correctly catch when the sound is
stopping when reaching eof, which make the state property correctly
work.

Co-authored-by: Mathieu Virbel <tito@Hubble.local>
(cherry picked from commit 3eadfd9fdeb4f0fb1f1a8e786e88952e0a8c42ec)